+++ monitorFunctions/monitorSeason.py
