--- v1.0 - originalScript/docker-PlexSonarrAutoDownloader/crontab
@@ -1,3 +1,0 @@
-#run python script every 30 minutes
-*/30 * * * * /usr/local/bin/python3 /app/PlexSonarrAutoDownloader.py > /proc/1/fd/1 2>/proc/1/fd/2
-
