--- v1.0 - originalScript/python-PlexSonarrAutoDownloader/crontab - examples.txt
@@ -1,3 +1,0 @@
-
-#run python script every 30 minutes
-*/30 * * * * /usr/bin/python3 /APPLICATION_DIRECTORY/PlexSonarrAutoDownloader.py > /proc/1/fd/1 2>/proc/1/fd/2
