--- v1.0 - originalScript/python-PlexSonarrAutoDownloader/requiredDependancies.txt
@@ -1,1 +1,0 @@
-python3 -m pip install requests plexapi
