--- v1.0 - originalScript/source.txt
@@ -1,1 +1,0 @@
-the ORIGINAL VERSION script is courtesy of reddit user /u/xPeacefulDreams
