--- v2.0&3.0 - production/source.txt
@@ -1,3 +1,0 @@
-the PRODUCTION VERSION script is courtesy of reddit user /u/Grtgignsky
-
-https://git.gignsky.com/
