--- v2.0&3.0 - production/v2.0 - noMonitorFunction/requiredDependancies.txt
@@ -1,1 +1,0 @@
-python3 -m pip install requests plexapi
